server directory refactor
